Compatibility with Thingworx 8.4
Instead of redefining the prototype as read-only, it is now a computed property whose setter has no effect. This is because as of Thingworx 8.4 the widget runtime runs in strict mode and would crash when attempting to modify the prototype property of the constructor. Now the assignment should fail silently as before.